+++ rotkehlchen/tests/unit/decoders/test_main.py
@@ -119,6 +119,7 @@ def test_decoders_initialization(ethereum_transaction_decoder: EthereumTransacti
         'Sky',
         'SocketBridgeDecoder',
         'Stakedao',
+        'SummerFi',
         'Sushiswap',
         'Thegraph',
         'Uniswapv1',
@@ -240,6 +241,7 @@ def test_decoders_initialization(ethereum_transaction_decoder: EthereumTransacti
         'uphold',
         'magpie',
         'merkl',
+        'summer-fi',
     }